requirements.txt
apiflask
flask-cors
pywebview
pyinstaller
